fix int/uint blunder
Patches applied:

 * bonzini@gnu.org--2004b/smalltalk--devo--2.2--patch-597
   fix int/uint blunder


git-archimport-id: bonzini@gnu.org--2004b/smalltalk--stable--2.3--patch-113

+++ libgst/dict.inl
@@ -578,7 +578,7 @@ uintptr_t
 scramble (uintptr_t x)
 {
 #if SIZEOF_OOP == 4
-  int y = ~x;
+  uintptr_t y = ~x;
   x += (y << 10) | (y >> 22);
   x += (x << 6)  | (x >> 26);
   x -= (x << 16) | (x >> 16);
